bus-util: allow root to connect to a user bus
Drop the check for peer creds from bus_connect_user_systemd(), instead
solely rely on system permissions to connect to the user manager.

This is OK, since the user directories /run/user/<uid> are created with
mode 0700, which protects the systemd/private socket.

Furthermore, the checks in bus_check_peercred() are written from a
server's point of view, it's awkward to use them from a client.

This makes it possible to use `systemctl --user` as root by explicitly
setting $XDG_RUNTIME_DIR to point to an existing user directory.

Before this commit:
  root# XDG_RUNTIME_DIR=/run/user/1000 systemctl --user status
  Failed to connect to bus: Operation not permitted

After this commit:
  root# XDG_RUNTIME_DIR=/run/user/1000 systemctl --user status
  State: running
  CGroup: /user.slice/user-1000.slice/user@1000.service
  ...