meson: generate version tag from git
$ build/systemctl --version
systemd 239-3555-g6178cbb5b5
+PAM +AUDIT +SELINUX +IMA -APPARMOR +SMACK +SYSVINIT +UTMP +LIBCRYPTSETUP +GCRYPT +GNUTLS +ACL +XZ +LZ4 +SECCOMP +BLKID +ELFUTILS +KMOD -IDN2 +IDN +PCRE2 default-hierarchy=hybrid
$ git tag v240 -m 'v240'
$ ninja -C build
ninja: Entering directory `build'
[76/76] Linking target fuzz-unit-file.
$ build/systemctl --version
systemd 240
+PAM +AUDIT +SELINUX +IMA -APPARMOR +SMACK +SYSVINIT +UTMP +LIBCRYPTSETUP +GCRYPT +GNUTLS +ACL +XZ +LZ4 +SECCOMP +BLKID +ELFUTILS +KMOD -IDN2 +IDN +PCRE2 default-hierarchy=hybrid

This is very useful during development, because a precise version string is
embedded in the build product and displayed during boot, so we don't have to
guess answers for questions like "did I just boot the latest version or the one
from before?".

This change creates an overhead for "noop" builds. On my laptop, 'ninja -C
build' that does nothing goes from 0.1 to 1.2 s. It would be nice to avoid
this, but I think that ~1s is still acceptable.

Fixes #7183.

The handling of pkgconfigdatadir=no, pkgconfiglibdir=no is simplified: the idea
was that those vars could be configured to 'no' to not install the .pc files.
When set to '', they would be built but not installed. This was inherited from
the autoconf build system. But having this level of control doesn't seem
necessary, since creating those files is very quick. Also, this was implemented
only for systemd.pc and not the other .pc files. Let's simplify things and only
skip installation if the target dir is configured as 'no'.

The implementation seems quite complex, but there are reasons:

Why save to a file?

It would be possible to generate the version tag using run_command(), but we
want to go through a file so that stuff gets rebuilt when this file changes. If
we just defined an variable in meson, ninja wouldn't know it needs to rebuild
things.

Why both vcs_tag() and custom_target()?

meson has vcs_tag, but it cannot be combined with configure_file(), because
vcs_tag requires a plain file as input. Similarly, combining them in the
opposite order also doesn't work for similar reasons, and would be wrong,
because the "configuration" step (which depends on configuration settings),
should happen before the "git describe" step.

Why use a custom pattern (with %%)?

configure_file() doesn't understand @@...@@ as a means to avoid substitution,
so using a different pattern seems simplest.